add category to an expense

+++ src/main/java/com/simpleexpenses/demo/service/impl/ExpenseServiceImpl.java
@@ -3,10 +3,13 @@ package com.simpleexpenses.demo.service.impl;
 import com.simpleexpenses.demo.dto.ExpenseDto;
 import com.simpleexpenses.demo.exceptions.AccessDeniedException;
 import com.simpleexpenses.demo.exceptions.EntityNotFoundException;
+import com.simpleexpenses.demo.model.entity.CategoryEntity;
 import com.simpleexpenses.demo.model.entity.ExpenseEntity;
 import com.simpleexpenses.demo.model.entity.ExpensesGroupEntity;
+import com.simpleexpenses.demo.repository.CategoryRepository;
 import com.simpleexpenses.demo.repository.ExpenseRepository;
 import com.simpleexpenses.demo.repository.ExpensesGroupRepository;
+import com.simpleexpenses.demo.service.CategoryService;
 import com.simpleexpenses.demo.service.ExpenseService;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
@@ -22,6 +25,7 @@ public class ExpenseServiceImpl implements ExpenseService {
 
     private final ExpensesGroupRepository expensesGroupRepository;
     private final ExpenseRepository expenseRepository;
+    private final CategoryRepository categoryRepository;
 
     @Override
     public ExpenseDto createExpense(ExpenseDto expenseDto) {
@@ -52,7 +56,7 @@ public class ExpenseServiceImpl implements ExpenseService {
         return storedExpenseDto;
     }
 
-    private void checkExpenseAcces(ExpenseEntity expenseEntity) {
+    private void checkExpenseAccess(ExpenseEntity expenseEntity) {
         ExpensesGroupEntity expensesGroup = this.expensesGroupRepository
                 .findByExpensesGroupId(expenseEntity.getExpensesGroup().getExpensesGroupId())
                 .orElseThrow(() -> new EntityNotFoundException("Expenses group doesn't exist."));
@@ -70,7 +74,7 @@ public class ExpenseServiceImpl implements ExpenseService {
         ExpenseEntity expenseEntity = this.expenseRepository.findByExpenseId(expenseId)
                 .orElseThrow(()  -> new EntityNotFoundException("Expense doesn't exist."));
 
-        checkExpenseAcces(expenseEntity);
+        checkExpenseAccess(expenseEntity);
 
         ModelMapper modelMapper = new ModelMapper();
 
@@ -96,9 +100,35 @@ public class ExpenseServiceImpl implements ExpenseService {
         ExpenseEntity expenseEntity = this.expenseRepository.findByExpenseId(expenseId)
                 .orElseThrow(()  -> new EntityNotFoundException("Expense doesn't exist."));
 
-        checkExpenseAcces(expenseEntity);
+        checkExpenseAccess(expenseEntity);
 
         this.expenseRepository.delete(expenseEntity);
+
+    }
+
+    @Override
+    public void addCategory(String categoryId, String expenseId) {
+
+        ExpenseEntity expenseEntity = this.expenseRepository
+                .findByExpenseId(expenseId)
+                .orElseThrow(() -> new EntityNotFoundException("Expense doesn't exist."));
+
+        checkExpenseAccess(expenseEntity);
+
+        CategoryEntity categoryEntity = this.categoryRepository
+                .findByCategoryId(categoryId)
+                .orElseThrow(() -> new EntityNotFoundException("Category doesn't exist."));
+
+        String userId = SecurityContextHolder.getContext().getAuthentication().getName();
+
+        if (!categoryEntity.getUserId().equals(userId)) {
+            throw new AccessDeniedException("You can't use this category.");
+        }
+
+        if (!expenseEntity.getCategories().contains(categoryEntity)) {
+            expenseEntity.getCategories().add(categoryEntity);
+            this.expenseRepository.save(expenseEntity);
+        }
 
     }
 }
